+++ test/bad-webconfig/server.js
@@ -1,4 +1,95 @@
-require('node-atlas').config({
+const NA = require("../../")();
+
+for (var i in NA) {
+	console.log("=====");
+	console.log(i);
+	console.log("=====");
+}
+
+/*require("../../").init({
     httpPort: 7777,
     webconfig: "webconfig.no-webconfig.json"
-}).init();
+}).start();*/
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+=====
+addCommonVariation
+=====
+=====
+addSpecificVariation
+=====
+=====
+newRender
+=====
+
+
+
+
+=====
+lessCompilation
+=====
+=====
+stylusCompilation
+=====
+=====
+cssCompilation
+=====
+=====
+cssMinification
+=====
+=====
+imgOptimization
+=====
+=====
+jsObfuscation
+=====
+
+
+
+
+
+
+
+
+## Disponible dès le début, mais pas dans l'API global ##
+
+init
+generated
+started
+created
+start
+run
+
+## Disponible dès le début et public
+extend
+clone
+forEach
+
+
+## Disponible à partir d'un certain point ##
+appLabels (devenir cliLabels)
+appLanguage (devenir cliLanguage / changer valeur par défaut ?)
+nodeatlasModulesRelativePath
+nodeatlasPath
+serverModulesRelativePath
+serverPath
+sessionStore
+webconfigName
+httpServer (express)
+server (http.createServer))